requirements.txt
flask
python-dotenv
requests
beautifulsoup4
python-dateutil
numpy
torch
xgboost
scikit-learn
catboost
